+++ src/Table/View/Cells/Actions/Dots.php
@@ -6,11 +6,22 @@ namespace Patrikjak\Utils\Table\View\Cells\Actions;
 
 use Illuminate\Contracts\View\View;
 use Illuminate\View\Component;
+use Patrikjak\Utils\Table\Dto\Cells\Actions\Item;
 
 class Dots extends Component
 {
-    public function __construct(public ?string $actionsDataAttributes = null)
-    {
+    /**
+     * @param array<Item> $inlineActions
+     * @param array<string> $hiddenInlineActionIds
+     * @param array<string, string|null> $inlineActionHrefs
+     */
+    public function __construct(
+        public ?string $actionsDataAttributes = null,
+        public array $inlineActions = [],
+        public array $hiddenInlineActionIds = [],
+        public bool $hasDropdownActions = false,
+        public array $inlineActionHrefs = [],
+    ) {
     }
 
     public function render(): View