+++ tests/Unit/Table/Http/Request/TableParametersRequestFilterTest.php
@@ -5,8 +5,6 @@ declare(strict_types=1);
 namespace Patrikjak\Utils\Tests\Unit\Table\Http\Request;
 
 use Carbon\CarbonInterface;
-use Illuminate\Contracts\Container\BindingResolutionException;
-use Illuminate\Cookie\CookieJar;
 use Orchestra\Testbench\TestCase;
 use Patrikjak\Utils\Common\Dto\Filter\DateFilterCriteria;
 use Patrikjak\Utils\Common\Dto\Filter\NumberFilterCriteria;
@@ -17,8 +15,6 @@ use Patrikjak\Utils\Table\Http\Requests\TableParametersRequest;
 class TableParametersRequestFilterTest extends TestCase
 {
     private const string TABLE_ID = 'table-id';
-
-    private CookieJar $cookieJar;
 
     public function testGetTableParametersFilterCriteriaFromRequest(): void
     {
@@ -61,28 +57,8 @@ class TableParametersRequestFilterTest extends TestCase
             ],
         ]);
 
-        $request->cookies->set(
-            self::TABLE_ID,
-            json_encode([
-                'page' => 5,
-                'pageSize' => 50,
-                'sortCriteria' => null,
-                'filterCriteria' => [
-                    'location' => [
-                        [
-                            'value' => 'New York',
-                            'type' => 'text',
-                            'operator' => 'contains',
-                        ],
-                    ],
-                ],
-            ]),
-        );
-
         $parameters = $request->getTableParameters(self::TABLE_ID);
 
-        $this->assertSame(5, $parameters->page);
-        $this->assertSame(50, $parameters->pageSize);
         $this->assertCount(5, $parameters->filterCriteria->filters);
 
         $this->assertInstanceOf(TextFilterCriteria::class, $parameters->filterCriteria->filters[0]);
@@ -112,67 +88,6 @@ class TableParametersRequestFilterTest extends TestCase
         $this->assertInstanceOf(NumberFilterCriteria::class, $parameters->filterCriteria->filters[4]);
     }
 
-    public function testGetTableParametersFilterCriteriaFromCookie(): void
-    {
-        $request = new TableParametersRequest();
-        $request->cookies->set(
-            self::TABLE_ID,
-            json_encode([
-                'page' => 5,
-                'pageSize' => 50,
-                'sortCriteria' => null,
-                'filterCriteria' => [
-                    [
-                        'column' => 'location',
-                        'value' => 'New York',
-                        'operator' => 'contains',
-                    ],
-                    [
-                        'column' => 'name',
-                        'value' => 'John',
-                        'type' => 'text',
-                    ],
-                    [
-                        'column' => 'color',
-                        'value' => 'red',
-                        'type' => 'select',
-                    ],
-                    [
-                        'column' => 'created_at',
-                        'from' => '2024-12-01',
-                        'to' => '2024-12-31',
-                        'type' => 'date',
-                    ],
-                    [
-                        'column' => 'random_number',
-                        'from' => 10,
-                        'type' => 'number',
-                    ],
-                ],
-            ]),
-        );
-
-        $parameters = $request->getTableParameters(self::TABLE_ID);
-
-        $this->assertSame(5, $parameters->page);
-        $this->assertSame(50, $parameters->pageSize);
-        $this->assertCount(3, $parameters->filterCriteria->filters);
-
-        $this->assertInstanceOf(SelectFilterCriteria::class, $parameters->filterCriteria->filters[0]);
-        $this->assertSame('color', $parameters->filterCriteria->filters[0]->column);
-        $this->assertSame('red', $parameters->filterCriteria->filters[0]->value);
-
-        $this->assertInstanceOf(DateFilterCriteria::class, $parameters->filterCriteria->filters[1]);
-        $this->assertSame('created_at', $parameters->filterCriteria->filters[1]->column);
-        $this->assertInstanceOf(CarbonInterface::class, $parameters->filterCriteria->filters[1]->from);
-        $this->assertInstanceOf(CarbonInterface::class, $parameters->filterCriteria->filters[1]->to);
-
-        $this->assertInstanceOf(NumberFilterCriteria::class, $parameters->filterCriteria->filters[2]);
-        $this->assertSame('random_number', $parameters->filterCriteria->filters[2]->column);
-        $this->assertSame(10.0, $parameters->filterCriteria->filters[2]->from);
-        $this->assertNull($parameters->filterCriteria->filters[2]->to);
-    }
-
     public function testGetTableParametersFilterCriteriaFromRequestWithInvalidType(): void
     {
         $request = new TableParametersRequest([
@@ -187,20 +102,8 @@ class TableParametersRequestFilterTest extends TestCase
             ],
         ]);
 
-        $request->cookies->set(
-            self::TABLE_ID,
-            json_encode([
-                'page' => 5,
-                'pageSize' => 50,
-                'sortCriteria' => null,
-                'filterCriteria' => null,
-            ]),
-        );
-
         $parameters = $request->getTableParameters(self::TABLE_ID);
 
-        $this->assertSame(5, $parameters->page);
-        $this->assertSame(50, $parameters->pageSize);
         $this->assertEmpty($parameters->filterCriteria->filters);
     }
 
@@ -210,72 +113,17 @@ class TableParametersRequestFilterTest extends TestCase
             'deleteFilters' => true,
         ]);
 
-        $request->cookies->set(
-            self::TABLE_ID,
-            json_encode([
-                'page' => 5,
-                'pageSize' => 50,
-                'sortCriteria' => null,
-                'filterCriteria' => [
-                    [
-                        'column' => 'location',
-                        'value' => 'New York',
-                        'type' => 'text',
-                        'operator' => 'contains',
-                    ],
-                    [
-                        'column' => 'name',
-                        'value' => 'John',
-                        'type' => 'text',
-                    ],
-                    [
-                        'column' => 'color',
-                        'value' => 'red',
-                        'type' => 'select',
-                    ],
-                    [
-                        'column' => 'created_at',
-                        'from' => '2024-12-01',
-                        'to' => '2024-12-31',
-                    ],
-                    [
-                        'column' => 'random_number',
-                        'from' => 10,
-                        'type' => 'number',
-                    ],
-                ],
-            ]),
-        );
+        $parameters = $request->getTableParameters(self::TABLE_ID);
+
+        $this->assertNull($parameters->filterCriteria);
+    }
+
+    public function testFilterCriteriaIsNullByDefault(): void
+    {
+        $request = new TableParametersRequest();
 
         $parameters = $request->getTableParameters(self::TABLE_ID);
 
-        $this->assertSame(5, $parameters->page);
-        $this->assertSame(50, $parameters->pageSize);
         $this->assertNull($parameters->filterCriteria);
-
-        $cookie = $this->cookieJar->getQueuedCookies()[0];
-
-        $this->assertNotNull($cookie);
-        $this->assertSame(self::TABLE_ID, $cookie->getName());
-        $this->assertSame(
-            json_encode([
-                'page' => 5,
-                'pageSize' => 50,
-                'sortCriteria' => null,
-                'filterCriteria' => null,
-                'searchQuery' => null,
-            ]),
-            $cookie->getValue(),
-        );
-    }
-
-    /**
-     * @throws BindingResolutionException
-     */
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        $this->cookieJar = $this->app->make(CookieJar::class);
     }
 }