+++ tests/Unit/Table/Http/Request/TableParametersRequestTest.php
@@ -4,21 +4,12 @@ declare(strict_types=1);
 
 namespace Patrikjak\Utils\Tests\Unit\Table\Http\Request;
 
-use Carbon\CarbonInterface;
-use Illuminate\Contracts\Container\BindingResolutionException;
-use Illuminate\Cookie\CookieJar;
 use Orchestra\Testbench\TestCase;
-use Patrikjak\Utils\Common\Dto\Filter\DateFilterCriteria;
-use Patrikjak\Utils\Common\Dto\Filter\NumberFilterCriteria;
-use Patrikjak\Utils\Common\Dto\Filter\SelectFilterCriteria;
-use Patrikjak\Utils\Common\Dto\Filter\TextFilterCriteria;
 use Patrikjak\Utils\Table\Http\Requests\TableParametersRequest;
 
 class TableParametersRequestTest extends TestCase
 {
     private const string TABLE_ID = 'table-id';
-
-    private CookieJar $cookieJar;
 
     public function testGetTableParametersDefault(): void
     {
@@ -28,7 +19,6 @@ class TableParametersRequestTest extends TestCase
 
         $this->assertSame(1, $parameters->page);
         $this->assertSame(10, $parameters->pageSize);
-        $this->assertSame([], $this->cookieJar->getQueuedCookies());
     }
 
     public function testGetTableParametersFromRequest(): void
@@ -42,60 +32,19 @@ class TableParametersRequestTest extends TestCase
 
         $this->assertSame(2, $parameters->page);
         $this->assertSame(20, $parameters->pageSize);
-
-        $cookie = $this->cookieJar->getQueuedCookies()[0];
-
-        $this->assertNotNull($cookie);
-        $this->assertSame(self::TABLE_ID, $cookie->getName());
-        $this->assertSame(
-            json_encode([
-                'page' => 2,
-                'pageSize' => 20,
-                'sortCriteria' => null,
-                'filterCriteria' => null,
-                'searchQuery' => null,
-            ]),
-            $cookie->getValue(),
-        );
-    }
-
-    public function testGetTableParametersFromCookie(): void
-    {
-        $request = new TableParametersRequest(cookies: []);
-        $request->cookies->set(self::TABLE_ID, json_encode(['page' => 3, 'pageSize' => 30, 'sortCriteria' => null]));
-
-        $parameters = $request->getTableParameters(self::TABLE_ID);
-
-        $this->assertSame(3, $parameters->page);
-        $this->assertSame(30, $parameters->pageSize);
     }
 
     public function testGetTableParametersPageFromRequest(): void
     {
         $request = new TableParametersRequest([
             'page' => 4,
+            'pageSize' => 50,
         ]);
-        $request->cookies->set(self::TABLE_ID, json_encode(['page' => 5, 'pageSize' => 50, 'sortCriteria' => null]));
 
         $parameters = $request->getTableParameters(self::TABLE_ID);
 
         $this->assertSame(4, $parameters->page);
         $this->assertSame(50, $parameters->pageSize);
-
-        $cookie = $this->cookieJar->getQueuedCookies()[0];
-
-        $this->assertNotNull($cookie);
-        $this->assertSame(self::TABLE_ID, $cookie->getName());
-        $this->assertSame(
-            json_encode([
-                'page' => 4,
-                'pageSize' => 50,
-                'sortCriteria' => null,
-                'filterCriteria' => null,
-                'searchQuery' => null,
-            ]),
-            $cookie->getValue(),
-        );
     }
 
     public function testGetTableParametersPageSizeFromRequest(): void
@@ -103,27 +52,11 @@ class TableParametersRequestTest extends TestCase
         $request = new TableParametersRequest([
             'pageSize' => 40,
         ]);
-        $request->cookies->set(self::TABLE_ID, json_encode(['page' => 5, 'pageSize' => 50, 'sortCriteria' => null]));
 
         $parameters = $request->getTableParameters(self::TABLE_ID);
 
-        $this->assertSame(5, $parameters->page);
+        $this->assertSame(1, $parameters->page);
         $this->assertSame(40, $parameters->pageSize);
-
-        $cookie = $this->cookieJar->getQueuedCookies()[0];
-
-        $this->assertNotNull($cookie);
-        $this->assertSame(self::TABLE_ID, $cookie->getName());
-        $this->assertSame(
-            json_encode([
-                'page' => 5,
-                'pageSize' => 40,
-                'sortCriteria' => null,
-                'filterCriteria' => null,
-                'searchQuery' => null,
-            ]),
-            $cookie->getValue(),
-        );
     }
 
     public function testGetTableParametersSortCriteriaFromRequest(): void
@@ -132,58 +65,20 @@ class TableParametersRequestTest extends TestCase
             'sort' => 'name',
             'order' => 'desc',
         ]);
-        $request->cookies->set(
-            self::TABLE_ID,
-            json_encode([
-                'page' => 5,
-                'pageSize' => 50,
-                'sortCriteria' => ['column' => 'location', 'order' => 'asc'],
-            ]),
-        );
 
         $parameters = $request->getTableParameters(self::TABLE_ID);
 
-        $this->assertSame(5, $parameters->page);
-        $this->assertSame(50, $parameters->pageSize);
         $this->assertSame('name', $parameters->sortCriteria->column);
         $this->assertSame('desc', $parameters->sortCriteria->order->value);
-
-        $cookie = $this->cookieJar->getQueuedCookies()[0];
-
-        $this->assertNotNull($cookie);
-        $this->assertSame(self::TABLE_ID, $cookie->getName());
-        $this->assertSame(
-            json_encode(
-                [
-                    'page' => 5,
-                    'pageSize' => 50,
-                    'sortCriteria' => ['column' => 'name', 'order' => 'desc'],
-                    'filterCriteria' => null,
-                    'searchQuery' => null,
-                ],
-            ),
-            $cookie->getValue(),
-        );
     }
 
-    public function testGetTableParametersSortCriteriaFromCookie(): void
+    public function testGetTableParametersSortCriteriaIsNullByDefault(): void
     {
         $request = new TableParametersRequest();
-        $request->cookies->set(
-            self::TABLE_ID,
-            json_encode([
-                'page' => 5,
-                'pageSize' => 50,
-                'sortCriteria' => ['column' => 'location', 'order' => 'asc'],
-            ]),
-        );
 
         $parameters = $request->getTableParameters(self::TABLE_ID);
 
-        $this->assertSame(5, $parameters->page);
-        $this->assertSame(50, $parameters->pageSize);
-        $this->assertSame('location', $parameters->sortCriteria->column);
-        $this->assertSame('asc', $parameters->sortCriteria->order->value);
+        $this->assertNull($parameters->sortCriteria);
     }
 
     public function testGetTableParametersWithDeleteSortCriteria(): void
@@ -194,44 +89,8 @@ class TableParametersRequestTest extends TestCase
             'deleteSort' => true,
         ]);
 
-        $request->cookies->set(
-            self::TABLE_ID,
-            json_encode([
-                'page' => 5,
-                'pageSize' => 50,
-                'sortCriteria' => ['column' => 'location', 'order' => 'asc'],
-            ]),
-        );
-
         $parameters = $request->getTableParameters(self::TABLE_ID);
 
-        $this->assertSame(5, $parameters->page);
-        $this->assertSame(50, $parameters->pageSize);
         $this->assertNull($parameters->sortCriteria);
-
-        $cookie = $this->cookieJar->getQueuedCookies()[0];
-
-        $this->assertNotNull($cookie);
-        $this->assertSame(self::TABLE_ID, $cookie->getName());
-        $this->assertSame(
-            json_encode([
-                'page' => 5,
-                'pageSize' => 50,
-                'sortCriteria' => null,
-                'filterCriteria' => null,
-                'searchQuery' => null,
-            ]),
-            $cookie->getValue(),
-        );
-    }
-
-    /**
-     * @throws BindingResolutionException
-     */
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        $this->cookieJar = $this->app->make(CookieJar::class);
     }
 }